refactor: Answer와 Question 클래스를 BaseEntity로 변경
- Answer와 Question 클래스가 BaseEntity를 상속받도록 수정하였습니다.
- id, createdDate, updatedDate 필드를 BaseEntity로 이동하여 중복 코드를 제거하였습니다.
- 생성자에서 id를 super로 호출하여 상위 클래스의 id를 설정하도록 변경하였습니다.
