+++ src/main/java/nextstep/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package nextstep.qna.domain;
 
+import nextstep.core.domain.BaseEntity;
 import nextstep.qna.CannotDeleteException;
 import nextstep.qna.NotFoundException;
 import nextstep.qna.UnAuthorizedException;
@@ -7,9 +8,7 @@ import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
 
-public class Answer {
-    private Long id;
-
+public class Answer extends BaseEntity {
     private NsUser writer;
 
     private Question question;
@@ -17,10 +16,6 @@ public class Answer {
     private String contents;
 
     private boolean deleted = false;
-
-    private LocalDateTime createdDate = LocalDateTime.now();
-
-    private LocalDateTime updatedDate;
 
     public Answer() {
     }
@@ -30,7 +25,7 @@ public class Answer {
     }
 
     public Answer(Long id, NsUser writer, Question question, String contents) {
-        this.id = id;
+        super(id);
         if(writer == null) {
             throw new UnAuthorizedException();
         }
@@ -42,10 +37,6 @@ public class Answer {
         this.writer = writer;
         this.question = question;
         this.contents = contents;
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public Answer setDeleted(boolean deleted) {
@@ -86,6 +77,6 @@ public class Answer {
 
     public DeleteHistory delete() {
         this.deleted = true;
-        return new DeleteHistory(ContentType.ANSWER, this.id, this.writer, LocalDateTime.now());
+        return new DeleteHistory(ContentType.ANSWER, getId(), this.writer, LocalDateTime.now());
     }
 }
